ML_2_GP_ITI/requirement.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
nltk>=3.6.0
umap-learn>=0.5.0
scipy>=1.7.0